Remove unused methods in JavabuilderFileManager interface
