+++ org-code-javabuilder/lib/src/main/java/dev/javabuilder/LocalMain.java
@@ -33,7 +33,7 @@ public class LocalMain {
             outputAdapter,
             ExecutionType.RUN,
             null,
-            new LocalFileManager(),
+            new LocalTempDirectoryManager(),
             new LifecycleNotifier());
     codeExecutionManager.execute();
   }
